refactor: loaderファイルを作成
