+++ app/.server/loaders/index.ts
@@ -1,0 +1,3 @@
+export * from './indexLoader'
+export * from './rankingLoader'
+export * from './trendLoader'
