+++ app/components/card/index.ts
@@ -2,4 +2,4 @@ import { AvatarCard } from './ranking-avatar-card'
 import { ClothCard } from './ranking-cloth-card'
 import { TopRankingCard } from './top-ranking-card'
 import { TopTrendCard } from './top-trend-card'
-export { AvatarCard, ClothCard, TopRankingCard ,TopTrendCard}
+export { AvatarCard, ClothCard, TopRankingCard, TopTrendCard }
